requirements.txt
gitdb==4.0.5
JPype1==1.2.0
